requirements.txt
asgiref==3.8.1
Django==4.2.17
django-cors-headers==4.6.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.12.2
python-dotenv==1.0.1
drf-spectacular==0.27.2
playwright>=1.40.0
requests>=2.31.0